% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cgmmetrics}
\alias{cgmmetrics}
\title{Calculate the metrics of CGM data.}
\usage{
cgmmetrics(
  inputdir,
  outputdir,
  useig = FALSE,
  diffnum = 1,
  threshold = 1,
  bthreshold = 3.9,
  athreshold = 10,
  interval = 15
)
}
\arguments{
\item{inputdir}{Path of input directory containing CGM files.}

\item{outputdir}{Path of output directory where metrics will be stored.}

\item{useig}{Logical. If TRUE the imputed data will be used.}

\item{diffnum}{The number of differencing.}

\item{threshold}{The threshold of sd used in mage calculation.}

\item{bthreshold}{The minimum glucose of tir range.}

\item{athreshold}{The maximum glucose of tir range.}

\item{interval}{The interval of CGM data.}
}
\description{
Calculate the metrics of CGM data.
}
