% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgmplot.R
\name{cgmplot}
\alias{cgmplot}
\title{Generate plots of CGM data.}
\usage{
cgmplot(
  inputdir,
  outputdir,
  useig = FALSE,
  markoutliers = TRUE,
  interval = 15,
  diffnum = 1,
  html = TRUE
)
}
\arguments{
\item{inputdir}{Path of input directory containing CGM files.}

\item{outputdir}{Path of output directory where plots will be stored.}

\item{useig}{Logical. If TRUE the imputed data will be used.}

\item{markoutliers}{Logical. If TRUE the outliers will be labeled in different colors.}

\item{interval}{The interval of CGM data.}

\item{diffnum}{The number of differencing.}

\item{html}{Logical.If TRUE the interactive plot will be exported, else the static pdf plot will exported.}
}
\description{
Generate plots of CGM data.
}
