% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfun.R
\name{qcfun}
\alias{qcfun}
\title{Quality Control for CGM data, include imputation and outlier detection}
\usage{
qcfun(
  cgmts,
  outlierdet = TRUE,
  interval = 15,
  imputation = FALSE,
  immethod = "linear",
  maxgap = 60,
  compeleteday = TRUE,
  removeday = FALSE,
  transunits = FALSE,
  removeflday = TRUE
)
}
\description{
Quality Control for CGM data, include imputation and outlier detection
}
